+++ bin/node/bench/src/generator.rs
@@ -18,11 +18,12 @@ use std::{collections::HashMap, sync::Arc};
 
 use kvdb::KeyValueDB;
 use node_primitives::Hash;
-use sp_trie::{DBValue, trie_types::TrieDBMut, TrieMut};
-use hash_db::{HashDB, AsHashDB, Prefix, Hasher as _};
+use sp_trie::{trie_types::TrieDBMut, TrieMut};
 
-type Hasher = sp_core::Blake2Hasher;
+use crate::simple_trie::SimpleTrie;
 
+/// Generate trie from given `key_values`.
+///
 /// Will fill your database `db` with trie data from `key_values` and
 /// return root.
 pub fn generate_trie(
@@ -37,9 +38,9 @@ pub fn generate_trie(
 			hex::decode("03170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c111314").expect("null key is valid"),
 			Some(vec![0]),
 		);
-		let mut trie_generator = TrieGenerator { db, overlay: &mut overlay };
+		let mut trie = SimpleTrie { db, overlay: &mut overlay };
 		{
-			let mut trie_db = TrieDBMut::new(&mut trie_generator, &mut root);
+			let mut trie_db = TrieDBMut::new(&mut trie, &mut root);
 
 			for (key, value) in key_values {
 				trie_db.insert(&key, &value).expect("trie insertion failed");
@@ -47,7 +48,7 @@ pub fn generate_trie(
 
 			trie_db.commit();
 		}
-		( trie_generator.db, overlay )
+		( trie.db, overlay )
 	};
 
 	let mut transaction = db.transaction();
@@ -60,48 +61,4 @@ pub fn generate_trie(
 	db.write(transaction).expect("Failed to write transaction");
 
 	root
-}
-
-/// Immutable generated trie database with root.
-struct TrieGenerator<'a> {
-	db: Arc<dyn KeyValueDB>,
-	overlay: &'a mut HashMap<Vec<u8>, Option<Vec<u8>>>,
-}
-
-impl<'a> AsHashDB<Hasher, DBValue> for TrieGenerator<'a> {
-	fn as_hash_db(&self) -> &dyn hash_db::HashDB<Hasher, DBValue> { &*self }
-
-	fn as_hash_db_mut<'b>(&'b mut self) -> &'b mut (dyn HashDB<Hasher, DBValue> + 'b) {
-		&mut *self
-	}
-}
-
-impl<'a> HashDB<Hasher, DBValue> for TrieGenerator<'a> {
-	fn get(&self, key: &Hash, prefix: Prefix) -> Option<DBValue> {
-		let key = sp_trie::prefixed_key::<Hasher>(key, prefix);
-		if let Some(value) = self.overlay.get(&key) {
-			return value.clone();
-		}
-		self.db.get(0, &key).expect("Database backend error")
-	}
-
-	fn contains(&self, hash: &Hash, prefix: Prefix) -> bool {
-		self.get(hash, prefix).is_some()
-	}
-
-	fn insert(&mut self, prefix: Prefix, value: &[u8]) -> Hash {
-		let key = Hasher::hash(value);
-		self.emplace(key, prefix, value.to_vec());
-		key
-	}
-
-	fn emplace(&mut self, key: Hash, prefix: Prefix, value: DBValue) {
-		let key = sp_trie::prefixed_key::<Hasher>(&key, prefix);
-		self.overlay.insert(key, Some(value));
-	}
-
-	fn remove(&mut self, key: &Hash, prefix: Prefix) {
-		let key = sp_trie::prefixed_key::<Hasher>(key, prefix);
-		self.overlay.insert(key, None);
-	}
 }
